webhook: Add first run of code
This includes all the code needed to test and run this.

It should be pointed out that deploying as is would require a docker image
that was not created with our usual templating processes. That needs to be
sorted out either by simply using this as is on the nodes that deploy it,
committing it to the registry directly, since this is not a user container
or similar.  This webhook can be configured to allow any registry
for its own deployment, which would allow it to be in a separate space.

Bug: T215678
Change-Id: I8028c876c75aca56e9a5a55aabd77b62bc054d27